#21: renamed the interface to not include the full path with the package. Moved the companion object closer to the beginning of the class.
