+++ app/src/main/java/so/codex/hawk/logging/LogcatFormatter.kt
@@ -1,46 +1,23 @@
 package so.codex.hawk.logging
 
 import android.util.Log
-import java.text.SimpleDateFormat
-import java.util.Locale
 
 /**
- * A class that implements [so.codex.hawk.logging.Formatter].
- * It is designed to format the original message before being written to the log.
+ * To use your formatter in [FileLoggingTree], you need it to implement this interface.
  */
-object LogcatFormatter : Formatter {
-    /**
-     * @property priorityPrefixes contains prefixes to indicate the type of log record.
-     */
-    private val priorityPrefixes = HashMap<Int, String>()
-
-    /**
-     * @property dateFormat [SimpleDateFormat] object to convert time from [Long] to format
-     *                      according to the pattern.
-     */
-    private val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS", Locale.US)
-
-    /**
-     * Initialization block. Prefixes are set for each log type.
-     */
-    init {
-        priorityPrefixes.apply {
-            put(Log.VERBOSE, "V/")
-            put(Log.DEBUG, "D/")
-            put(Log.INFO, "I/")
-            put(Log.WARN, "W/")
-            put(Log.ERROR, "E/")
-            put(Log.ASSERT, "WTF/")
-        }
-    }
-
+interface LogcatFormatter {
     /**
      * The method of formatting the original message before writing it to the log.
-     * @see so.codex.hawk.logging.Formatter.format
+     *
+     * @param priority Log record priority. Depends on the method used when writing to the log.
+     * @see Log
+     *
+     * @param tag Used to identify the source of a log message. It usually identifies the class
+     *            or activity where the log call occurs.
+     *
+     * @param message contains the message sent to the log.
+     *
+     * @return processed message that will be written to the log.
      */
-    override fun format(priority: Int, tag: String?, message: String): String {
-        val prio = priorityPrefixes[priority] ?: ""
-        val time = dateFormat.format(System.currentTimeMillis())
-        return "[$time] $prio${tag ?: ""}: $message\n"
-    }
+    fun format(priority: Int, tag: String?, message: String): String
 }
